Time: 16 ms (86.41%), Space: 24.2 MB (37.27%) - LeetHub

+++ 2042-maximum-product-difference-between-two-pairs/2042-maximum-product-difference-between-two-pairs.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int maxProductDifference(vector<int>& nums) {
         sort(nums.begin(), nums.end());
-        int numsSize = nums.size();
+        unsigned short numsSize = nums.size();
 
         return (nums[numsSize - 1] * nums[numsSize - 2]) - (nums[0] * nums[1]);
     }
